Moved manager empty state to separate file to reuse it in Cypress tests
