+++ app/reducers/managerEmptyState.js
@@ -1,0 +1,16 @@
+export default {
+    auth: {
+        role: null,
+        groupSystemRoles: {},
+        tenantsRoles: {}
+    },
+    clusterStatus: {},
+    err: null,
+    isLoggingIn: false,
+    lastUpdated: null,
+    license: {},
+    permissions: {},
+    roles: [],
+    tenants: {},
+    version: {}
+};
